+++ tests/Integration/Invalidation/CancelInvalidationTest.php
@@ -4,6 +4,7 @@ namespace Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Invalidation;
 
 use FOS\HttpCacheBundle\CacheManager;
 use Neusta\Pimcore\HttpCacheBundle\Element\ElementInvalidationEvent;
+use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\ArrangeCacheTest;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestAssetFactory;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestDocumentFactory;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestObjectFactory;
@@ -16,6 +17,7 @@ use Prophecy\Prophecy\ObjectProphecy;
 
 final class CancelInvalidationTest extends ConfigurableKernelTestCase
 {
+    use ArrangeCacheTest;
     use ProphecyTrait;
     use ResetDatabase;
 
@@ -31,8 +33,6 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
             ElementInvalidationEvent::class,
             static fn ($event) => $event->cancel = true,
         );
-
-        self::getContainer()->get('neusta_pimcore_http_cache.cache_scope')->enable();
     }
 
     /**
@@ -45,7 +45,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_object_update(): void
     {
-        $object = TestObjectFactory::simpleObject()->save();
+        $object = self::arrange(static fn () => TestObjectFactory::simpleObject()->save());
 
         $object->setContent('Updated test content')->save();
 
@@ -62,7 +62,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_document_update(): void
     {
-        $document = TestDocumentFactory::simplePage()->save();
+        $document = self::arrange(static fn () => TestDocumentFactory::simplePage()->save());
 
         $document->setKey('updated_test_document_page')->save();
 
@@ -79,7 +79,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_asset_update(): void
     {
-        $asset = TestAssetFactory::simpleAsset()->save();
+        $asset = self::arrange(static fn () => TestAssetFactory::simpleAsset()->save());
 
         $asset->setData('Updated test content')->save();
 
@@ -96,7 +96,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_object_delete(): void
     {
-        $object = TestObjectFactory::simpleObject()->save();
+        $object = self::arrange(static fn () => TestObjectFactory::simpleObject()->save());
 
         $object->delete();
 
@@ -113,7 +113,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_document_delete(): void
     {
-        $document = TestDocumentFactory::simplePage()->save();
+        $document = self::arrange(static fn () => TestDocumentFactory::simplePage()->save());
 
         $document->delete();
 
@@ -130,7 +130,7 @@ final class CancelInvalidationTest extends ConfigurableKernelTestCase
     ])]
     public function cancel_invalidation_on_asset_delete(): void
     {
-        $asset = TestAssetFactory::simpleAsset()->save();
+        $asset = self::arrange(static fn () => TestAssetFactory::simpleAsset()->save());
 
         $asset->delete();