+++ tests/Integration/Invalidation/InvalidateObjectTest.php
@@ -3,6 +3,7 @@
 namespace Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Invalidation;
 
 use FOS\HttpCacheBundle\CacheManager;
+use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\ArrangeCacheTest;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestObjectFactory;
 use Neusta\Pimcore\TestingFramework\Database\ResetDatabase;
 use Neusta\Pimcore\TestingFramework\Test\Attribute\ConfigureExtension;
@@ -15,6 +16,7 @@ use Prophecy\Prophecy\ObjectProphecy;
 
 final class InvalidateObjectTest extends ConfigurableKernelTestCase
 {
+    use ArrangeCacheTest;
     use ProphecyTrait;
     use ResetDatabase;
 
@@ -33,11 +35,11 @@ final class InvalidateObjectTest extends ConfigurableKernelTestCase
         $this->cacheManager->invalidateTags(Argument::any())->willReturn($this->cacheManager->reveal());
         self::getContainer()->set('fos_http_cache.cache_manager', $this->cacheManager->reveal());
 
-        $this->object = TestObjectFactory::simpleObject()->save();
-        $this->folder = TestObjectFactory::simpleFolder()->save();
-        $this->variant = TestObjectFactory::simpleVariant()->save();
-
-        self::getContainer()->get('neusta_pimcore_http_cache.cache_scope')->enable();
+        self::arrange(function (): void {
+            $this->object = TestObjectFactory::simpleObject()->save();
+            $this->folder = TestObjectFactory::simpleFolder()->save();
+            $this->variant = TestObjectFactory::simpleVariant()->save();
+        });
     }
 
     /**