+++ tests/Integration/Tagging/TagAdditionalTagTest.php
@@ -7,6 +7,7 @@ use Neusta\Pimcore\HttpCacheBundle\Cache\CacheType\CustomCacheType;
 use Neusta\Pimcore\HttpCacheBundle\Cache\CacheType\ElementCacheType;
 use Neusta\Pimcore\HttpCacheBundle\Element\ElementTaggingEvent;
 use Neusta\Pimcore\HttpCacheBundle\Element\ElementType;
+use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\ArrangeCacheTest;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestAssetFactory;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestDocumentFactory;
 use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\Helpers\TestObjectFactory;
@@ -21,6 +22,7 @@ use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 #[ConfigureRoute(__DIR__ . '/../Fixtures/get_object_route.php')]
 final class TagAdditionalTagTest extends ConfigurableWebTestcase
 {
+    use ArrangeCacheTest;
     use ResetDatabase;
 
     private KernelBrowser $client;
@@ -40,8 +42,10 @@ final class TagAdditionalTagTest extends ConfigurableWebTestcase
     ])]
     public function response_is_tagged_with_additional_tag_when_asset_is_loaded(): void
     {
-        TestAssetFactory::simpleAsset()->save();
-        TestAssetFactory::simpleImage()->save();
+        self::arrange(static function (): void {
+            TestAssetFactory::simpleAsset()->save();
+            TestAssetFactory::simpleImage()->save();
+        });
 
         self::getContainer()->get('event_dispatcher')->addListener(
             ElementTaggingEvent::class,
@@ -70,8 +74,10 @@ final class TagAdditionalTagTest extends ConfigurableWebTestcase
     ])]
     public function response_is_tagged_with_additional_tag_when_document_is_loaded(): void
     {
-        TestDocumentFactory::simplePage()->save();
-        TestDocumentFactory::simpleSnippet()->save();
+        self::arrange(static function (): void {
+            TestDocumentFactory::simplePage()->save();
+            TestDocumentFactory::simpleSnippet()->save();
+        });
 
         self::getContainer()->get('event_dispatcher')->addListener(
             ElementTaggingEvent::class,
@@ -100,8 +106,10 @@ final class TagAdditionalTagTest extends ConfigurableWebTestcase
     ])]
     public function response_is_tagged_with_additional_tag_when_object_is_loaded(): void
     {
-        TestObjectFactory::simpleObject()->save();
-        TestObjectFactory::simpleVariant()->save();
+        self::arrange(static function (): void {
+            TestObjectFactory::simpleObject()->save();
+            TestObjectFactory::simpleVariant()->save();
+        });
 
         self::getContainer()->get('event_dispatcher')->addListener(
             ElementTaggingEvent::class,
@@ -133,7 +141,7 @@ final class TagAdditionalTagTest extends ConfigurableWebTestcase
     ])]
     public function response_is_tagged_with_custom_tag_when_element_is_loaded(): void
     {
-        TestObjectFactory::simpleObject()->save();
+        self::arrange(static fn () => TestObjectFactory::simpleObject()->save());
 
         self::getContainer()->get('event_dispatcher')->addListener(
             ElementTaggingEvent::class,